refactor: move add bookmark and manager to the domain
