+++ packages/views/screens/router/routesMap.js
@@ -1,7 +1,7 @@
-import AddBookmark from '@screens/managers/bookmarks/addBookmark';
+import AddBookmark from '@bookmark/manager/AddBookmarkManager';
 import BlockDetails from '@screens/managers/monitor/blockDetails';
 import Blocks from '@screens/managers/monitor/blocks';
-import Bookmarks from '@screens/managers/bookmarks/list';
+import Bookmarks from '@bookmark/manager/BookmarkListManager';
 import Dashboard from '@screens/managers/dashboard';
 import DelegatesMonitor from '@screens/managers/monitor/delegates';
 import DelegatesPerformanceModal from '@screens/managers/wallet/delegatePerformanceModal';
