+++ src/modules/bookmark/components/AddBookmark/AddBookmark.css
@@ -1,4 +1,4 @@
-@import '../../../../../../setup/react/app/mixins.css';
+@import '../../../../../setup/react/app/mixins.css';
 
 .wrapper {
   display: flex;
