+++ src/modules/bookmark/components/AddBookmark/index.js
@@ -1,0 +1,1 @@
+export { default } from './AddBookmark';
