+++ src/modules/bookmark/components/BookmarksListModal/BookmarkListModal.js
@@ -1,15 +1,15 @@
 import React from 'react';
 import { withRouter } from 'react-router';
 import { removeSearchParamsFromUrl, addSearchParamsToUrl } from 'src/utils/searchParams';
-import List from '@bookmark/components/BookmarksList';
-import ModalWrapper from '../modalWrapper';
-import styles from '../bookmarks.css';
+import BookmarksList from '@bookmark/components/BookmarksList';
+import ModalWrapper from '@bookmark/components/BookmarksListModal/BookmarkModalWrapper';
+import styles from '@bookmark/components/BookmarksListModal/BookmarkListModal.css';
 
 const Bookmarks = ({
   bookmarks, token, t, bookmarkRemoved, bookmarkUpdated, history,
 }) => (
   <ModalWrapper>
-    <List
+    <BookmarksList
       enableFilter
       isEditable
       bookmarks={bookmarks}
