+++ src/modules/bookmark/components/BookmarksListModal/BookmarkListModal.test.js
@@ -1,7 +1,7 @@
 import { tokenMap } from '@token/configuration/tokens';
 import { mountWithRouter } from '@common/utilities/testHelpers';
 import bookmarks from '@tests/constants/bookmarks';
-import BookmarkListModal from './modal';
+import BookmarkListModal from './BookmarkListModal';
 
 describe('BookmarkListModal', () => {
   let wrapper;
