+++ src/modules/bookmark/components/EmptyState/index.js
@@ -1,1 +1,1 @@
-export default from './EmptyState';
+export { default } from './EmptyState';
