+++ src/modules/bookmark/manager/AddBookmarkManager.js
@@ -7,7 +7,7 @@ import { bookmarkAdded, bookmarkUpdated, bookmarkRemoved } from '@common/store/a
 import { getAccount } from '@wallet/utilities/api';
 import withData from '@common/utilities/withData';
 import { selectSearchParamValue } from 'src/utils/searchParams';
-import AddBookmark from './addBookmark';
+import AddBookmark from '../components/AddBookmark';
 
 const mapStateToProps = state => ({
   bookmarks: state.bookmarks,
