+++ src/modules/bookmark/manager/BookmarkListManager.js
@@ -2,7 +2,7 @@
 import { connect } from 'react-redux';
 import { withTranslation } from 'react-i18next';
 import { bookmarkRemoved, bookmarkUpdated } from '@common/store/actions';
-import BookmarkListModal from './modal';
+import BookmarkListModal from '@bookmark/components/BookmarksListModal';
 
 const mapStateToProps = state => ({
   bookmarks: state.bookmarks,
